+++ lib/twilio-ruby/rest/sync/v1/service/document.rb
@@ -30,9 +30,14 @@ module Twilio
             # Request is executed immediately.
             # @param [String] unique_name The unique_name
             # @param [Hash] data The data
+            # @param [String] ttl The ttl
             # @return [DocumentInstance] Newly created DocumentInstance
-            def create(unique_name: :unset, data: :unset)
-              data = Twilio::Values.of({'UniqueName' => unique_name, 'Data' => Twilio.serialize_object(data)})
+            def create(unique_name: :unset, data: :unset, ttl: :unset)
+              data = Twilio::Values.of({
+                  'UniqueName' => unique_name,
+                  'Data' => Twilio.serialize_object(data),
+                  'Ttl' => ttl,
+              })
 
               payload = @version.create(
                   'POST',
@@ -208,9 +213,10 @@ module Twilio
             ##
             # Update the DocumentInstance
             # @param [Hash] data The data
+            # @param [String] ttl The ttl
             # @return [DocumentInstance] Updated DocumentInstance
-            def update(data: nil)
-              data = Twilio::Values.of({'Data' => Twilio.serialize_object(data)})
+            def update(data: :unset, ttl: :unset)
+              data = Twilio::Values.of({'Data' => Twilio.serialize_object(data), 'Ttl' => ttl})
 
               payload = @version.update(
                   'POST',
@@ -274,6 +280,7 @@ module Twilio
                   'links' => payload['links'],
                   'revision' => payload['revision'],
                   'data' => payload['data'],
+                  'date_expires' => Twilio.deserialize_iso8601_datetime(payload['date_expires']),
                   'date_created' => Twilio.deserialize_iso8601_datetime(payload['date_created']),
                   'date_updated' => Twilio.deserialize_iso8601_datetime(payload['date_updated']),
                   'created_by' => payload['created_by'],
@@ -344,6 +351,12 @@ module Twilio
             end
 
             ##
+            # @return [Time] The date_expires
+            def date_expires
+              @properties['date_expires']
+            end
+
+            ##
             # @return [Time] The date_created
             def date_created
               @properties['date_created']
@@ -378,9 +391,10 @@ module Twilio
             ##
             # Update the DocumentInstance
             # @param [Hash] data The data
+            # @param [String] ttl The ttl
             # @return [DocumentInstance] Updated DocumentInstance
-            def update(data: nil)
-              context.update(data: data)
+            def update(data: :unset, ttl: :unset)
+              context.update(data: data, ttl: ttl)
             end
 
             ##