+++ lib/twilio-ruby/rest/sync/v1/service/sync_list/sync_list_item.rb
@@ -31,9 +31,10 @@ module Twilio
               # Retrieve a single page of SyncListItemInstance records from the API.
               # Request is executed immediately.
               # @param [Hash] data The data
+              # @param [String] ttl The ttl
               # @return [SyncListItemInstance] Newly created SyncListItemInstance
-              def create(data: nil)
-                data = Twilio::Values.of({'Data' => Twilio.serialize_object(data)})
+              def create(data: nil, ttl: :unset)
+                data = Twilio::Values.of({'Data' => Twilio.serialize_object(data), 'Ttl' => ttl})
 
                 payload = @version.create(
                     'POST',
@@ -235,9 +236,10 @@ module Twilio
               ##
               # Update the SyncListItemInstance
               # @param [Hash] data The data
+              # @param [String] ttl The ttl
               # @return [SyncListItemInstance] Updated SyncListItemInstance
-              def update(data: nil)
-                data = Twilio::Values.of({'Data' => Twilio.serialize_object(data)})
+              def update(data: :unset, ttl: :unset)
+                data = Twilio::Values.of({'Data' => Twilio.serialize_object(data), 'Ttl' => ttl})
 
                 payload = @version.update(
                     'POST',
@@ -285,6 +287,7 @@ module Twilio
                     'url' => payload['url'],
                     'revision' => payload['revision'],
                     'data' => payload['data'],
+                    'date_expires' => Twilio.deserialize_iso8601_datetime(payload['date_expires']),
                     'date_created' => Twilio.deserialize_iso8601_datetime(payload['date_created']),
                     'date_updated' => Twilio.deserialize_iso8601_datetime(payload['date_updated']),
                     'created_by' => payload['created_by'],
@@ -358,6 +361,12 @@ module Twilio
               end
 
               ##
+              # @return [Time] The date_expires
+              def date_expires
+                @properties['date_expires']
+              end
+
+              ##
               # @return [Time] The date_created
               def date_created
                 @properties['date_created']
@@ -392,9 +401,10 @@ module Twilio
               ##
               # Update the SyncListItemInstance
               # @param [Hash] data The data
+              # @param [String] ttl The ttl
               # @return [SyncListItemInstance] Updated SyncListItemInstance
-              def update(data: nil)
-                context.update(data: data)
+              def update(data: :unset, ttl: :unset)
+                context.update(data: data, ttl: ttl)
               end
 
               ##