+++ lib/twilio-ruby/rest/video/v1/room/room_participant/room_participant_published_track.rb
@@ -9,7 +9,7 @@ module Twilio
     class Video < Domain
       class V1 < Version
         class RoomContext < InstanceContext
-          class RoomParticipantContext < InstanceContext
+          class ParticipantContext < InstanceContext
             class PublishedTrackList < ListResource
               ##
               # Initialize the PublishedTrackList