+++ spec/integration/sync/v1/service/sync_map/sync_map_item_spec.rb
@@ -32,6 +32,7 @@ describe 'SyncMapItem' do
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "created_by": "created_by",
           "data": {},
+          "date_expires": "2015-07-30T21:00:00Z",
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z",
           "key": "key",
@@ -106,6 +107,7 @@ describe 'SyncMapItem' do
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "created_by": "created_by",
           "data": {},
+          "date_expires": "2015-07-30T21:00:00Z",
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z",
           "key": "key",
@@ -177,6 +179,7 @@ describe 'SyncMapItem' do
                   "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "created_by": "created_by",
                   "data": {},
+                  "date_expires": "2015-07-30T21:00:00Z",
                   "date_created": "2015-07-30T20:00:00Z",
                   "date_updated": "2015-07-30T20:00:00Z",
                   "key": "key",
@@ -212,15 +215,14 @@ describe 'SyncMapItem' do
     expect {
       @client.sync.v1.services("ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
                      .sync_maps("MPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
-                     .sync_map_items("key").update(data: "{}")
+                     .sync_map_items("key").update()
     }.to raise_exception(Twilio::REST::TwilioError)
 
-    values = {'Data' => Twilio.serialize_object("{}")}
+    values = {}
     expect(
     @holodeck.has_request?(Holodeck::Request.new(
         method: 'post',
         url: 'https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Maps/MPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Items/key',
-        data: values,
     ))).to eq(true)
   end
 
@@ -232,6 +234,7 @@ describe 'SyncMapItem' do
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "created_by": "created_by",
           "data": {},
+          "date_expires": "2015-07-30T21:00:00Z",
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z",
           "key": "key",
@@ -245,7 +248,7 @@ describe 'SyncMapItem' do
 
     actual = @client.sync.v1.services("ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
                             .sync_maps("MPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
-                            .sync_map_items("key").update(data: "{}")
+                            .sync_map_items("key").update()
 
     expect(actual).to_not eq(nil)
   end