+++ spec/integration/video/v1/recording_spec.rb
@@ -41,6 +41,8 @@ describe 'Recording' do
           "grouping_sids": {
               "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
           },
+          "media_external_location": "https://my-super-duper-bucket.s3.amazonaws.com/my/path/",
+          "encryption_key": "public_key",
           "links": {
               "media": "https://video.twilio.com/v1/Recordings/RTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Media"
           }
@@ -113,6 +115,8 @@ describe 'Recording' do
                       "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                       "participant_sid": "PAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                   },
+                  "media_external_location": "https://my-super-duper-bucket.s3.amazonaws.com/my/path/",
+                  "encryption_key": "public_key",
                   "url": "https://video.twilio.com/v1/Recordings/RTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "links": {
                       "media": "https://video.twilio.com/v1/Recordings/RTaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Media"